+++ src/components/auth/sign-in/sign-in.component.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, SyntheticEvent, useState } from 'react';
-import { auth, signInWithGoogle } from '../../firebase/firebase.utils';
-import CustomButton from '../custom-button/custom-button.component';
-import FormInput from '../form-input/form-input.component';
+import { auth, signInWithGoogle } from '../../../firebase/firebase.utils';
+import CustomButton from '../../shared/base/custom-button/custom-button.component';
+import FormInput from '../../shared/form/form-input/form-input.component';
 import './sign-in.styles.scss';
 
 export type SignInCredentials = {
