+++ src/components/auth/sign-up/sign-up.styles.scss
