+++ src/components/cart/cart-dropdown/cart-dropdown.component.tsx
@@ -1,7 +1,7 @@
 import { useSelector } from 'react-redux';
-import { selectCartItems } from '../../redux/cart/cart.selectors';
+import { selectCartItems } from '../../../redux/cart/cart.selectors';
 import CartItem from '../cart-item/cart-item.component';
-import CustomButton from '../custom-button/custom-button.component';
+import CustomButton from '../../shared/base/custom-button/custom-button.component';
 import './cart-dropdown.styles.scss';
 
 const CartDropdown = () => {
