+++ src/components/cart/cart-icon/cart-icon.styles.scss
