+++ src/components/collection/collection-preview/collection-preview.component.tsx
@@ -1,4 +1,4 @@
-import { Item } from '../../pages/shop/shop.component';
+import { Item } from '../../../pages/shop/shop.component';
 import CollectionItem from '../collection-item/collection-item.component';
 import './collection-preview.styles.scss';
