+++ src/components/directories/directory/directory.component.tsx
