+++ src/components/directories/directory/directory.styles.scss
