+++ src/components/shared/base/custom-button/custom-button.styles.scss
