+++ src/pages/sign-in-up/sign-in-up.component.tsx
@@ -1,5 +1,5 @@
-import SignIn from '../../components/sign-in/sign-in.component';
-import SignUp from '../../components/sign-up/sign-up.component';
+import SignIn from '../../components/auth/sign-in/sign-in.component';
+import SignUp from '../../components/auth/sign-up/sign-up.component';
 import './sign-in-up.styles.scss';
 
 const SignInUpPage = () => (
