+++ src/redux/cart/cart.selectors.ts
@@ -8,6 +8,11 @@ export const selectCartItems = createSelector(
 	(cart) => cart.cartItems
 );
 
+export const selectCartHidden = createSelector(
+	[selectCart],
+	(cart) => cart.isDropdownHidden
+);
+
 export const selectCartItemsCount = createSelector(
 	[selectCartItems],
 	(cartItems) =>
